:sparkles: add borax.htmls module
